Meta: Reorganize file structure around helpers
And add stub for ft_getcwd to be built
