+++ include/helpers_getcwd.hpp
@@ -1,0 +1,8 @@
+#ifndef HELPERS_GETCWD_HPP
+#define HELPERS_GETCWD_HPP
+
+#include <string>
+
+std::string ft_getcwd();
+
+#endif  // HELPER_GETCWD_HPP
